Amend the previous cold start fix.
Note: by "cold start" I mean that when Xwindows starts rox fails to find
/usr/share/backgrounds (for instance because it's incorrectly symlinked
to a non-existing directory).  Hence rox resets the background to a dull
default grey color. Now fix path /usr/share/backgrounds so rox can find
it and start the wallpaper application, which is now in a "cold start"
state.