Fix some issues with 'Clear' button.
Sometimes 'Clear' didn't remove the current preview image or the cached
copy of the current preview image or the cached copy of the actual
background image. Fixed.
In Stretch mode, when 'Clear'ing the actual background image and
immediately 'Apply'ing it again, the status bar showed just "Stretched"
instead of "Stretched, Cached". Fixed.